1、完善字典、字典类型

+++ src/main/java/com/shayne/controller/UserController.java
@@ -20,6 +20,7 @@ import com.shayne.domain.User;
 import com.shayne.domain.vo.ApiResult;
 import com.shayne.domain.vo.PageData;
 import com.shayne.domain.vo.PageImpl;
+import com.shayne.domain.vo.ResetPasswordUserVo;
 import com.shayne.domain.vo.UserRoleVo;
 import com.shayne.service.UserService;
 
@@ -36,47 +37,123 @@ public class UserController extends BaseController {
     private UserService userService;
     
     /**
-     * 新增用户
-     * @param user
-     * @return
-     * String
+     * 新增/更新
+     * @param User user
+     * @return ApiResult<String>
      */
     @PostMapping
     public ApiResult<String> save(@RequestBody User user) {
+    	ApiResult<String> result = new ApiResult<String>();
     	//用户信息加密
     	String msg = "SUCCESS";
+    	result.setMsg(msg);
+    	result.setResult(true);
     	
     	if(null == user) {
     		msg = "用户信息为空！";
-    		ApiResult.fail(msg);
+    		result.setMsg(msg);
+    		result.setResult(false);
     	}
-    	user.setCreateTime(new Date());
+    	// 写入创建时间
+    	if(null == user.getCreateTime()) {
+    		user.setCreateTime(new Date());
+    	}
         try {
 			user = super.encrypt(user);
 		} catch (Exception e) {
-			logger.error("用户密码加密发生异常：");
 			msg = e.getMessage();
+			result.setMsg(msg);
+			result.setResult(false);
+			logger.error("用户密码加密发生异常：");
 			e.printStackTrace();
-			return ApiResult.fail(msg);
 		}
         
         //用户信息保存
         try {
 			user = userService.save(user);
-			return new ApiResult<>(msg);
 		} catch (Exception e) {
-			logger.error("保存用户发生异常：");
 			msg = e.getMessage();
+			result.setMsg(msg);
+			result.setResult(false);
+			logger.error("保存用户发生异常：");
 			e.printStackTrace();
 		}
-        return ApiResult.fail(msg);
+        return result;
     }
     
     /**
-     * 删除用户
-     * @param sysUser
-     * @return
-     * String
+     * 重置密码
+     * @param User user
+     * @return ApiResult<String>
+     */
+    @PostMapping(value = "reset")
+    public ApiResult<String> reset(@RequestBody ResetPasswordUserVo resetPasswordUserVo) {
+    	ApiResult<String> result = new ApiResult<String>();
+    	//用户信息加密
+    	String msg = "SUCCESS";
+    	result.setMsg(msg);
+    	result.setResult(true);
+    	
+    	// 入参为空
+    	if(null == resetPasswordUserVo) {
+    		msg = "密码不能为空！";
+    		result.setMsg(msg);
+    		result.setResult(false);
+    		return result;
+    	}
+    	
+    	// 判断用户是否存在
+    	Long userId = resetPasswordUserVo.getId();
+    	User user = userService.find(userId);
+    	if(null == user) {
+    		msg = "用户不存在！";
+    		result.setMsg(msg);
+    		result.setResult(false);
+    		return result;
+    	}
+    	
+    	// 判断旧密码是否正确
+    	String salt = user.getSalt();
+    	String oldUserPassWord = user.getPassword();
+    	String oldPassWord = resetPasswordUserVo.getOldPassword();
+    	String oldEncryptPassWord = md5Util.encrypt(oldPassWord, salt);
+    	if(!StringUtils.equalsIgnoreCase(oldUserPassWord, oldEncryptPassWord)) {
+    		msg = "旧密码不正确！";
+    		result.setMsg(msg);
+    		result.setResult(false);
+    		return result;
+    	}
+    	
+    	// 判断用户密码是否和历史相同
+    	String newPassWord = resetPasswordUserVo.getNewPassword();
+    	String encryptNewPassword = md5Util.encrypt(newPassWord, salt);
+    	if(StringUtils.equalsIgnoreCase(encryptNewPassword, oldUserPassWord)) {
+    		msg = "新密码不能和旧密码相同！";
+    		result.setMsg(msg);
+    		result.setResult(false);
+    		return result;
+    	}
+    	
+    	// 写入新密码
+    	user.setPassword(encryptNewPassword);
+        
+        //用户信息保存
+        try {
+			user = userService.save(user);
+		} catch (Exception e) {
+			msg = e.getMessage();
+			result.setMsg(msg);
+			result.setResult(false);
+			logger.error("更新用户密码发生系统异常：");
+			e.printStackTrace();
+		}
+        return result;
+    }
+    
+    /**
+     * 批量删除
+     * @param String ids
+     * @return ApiResult<String>
      */
     @DeleteMapping
     public ApiResult<String> delete(@RequestParam(value="id")  String ids) {
@@ -92,7 +169,14 @@ public class UserController extends BaseController {
 			}
 		}
 		try {
-			userService.delete(idsSet);
+			boolean hasRole = userService.hasRole(idsSet);
+			if(hasRole) {
+				msg = "用户有授权角色，请先清除授权再删除";
+				result.setMsg(msg);
+				result.setResult(false);
+			}else {
+				userService.delete(idsSet);
+			}
 		} catch (Exception e) {
 			result.setResult(false);
 			msg = e.getMessage();
@@ -104,9 +188,8 @@ public class UserController extends BaseController {
     
     /**
      * 查询用户
-     * @param id
-     * @return
-     * User
+     * @param Long id
+     * @return ApiResult<User>
      */
     @GetMapping
     public ApiResult<User> get(@RequestParam(value="id") Long id) {
@@ -116,21 +199,22 @@ public class UserController extends BaseController {
 			user = userService.find(id);
 			result.setData(user);
 			result.setResult(true);
-			return result;
 		} catch (Exception e) {
 			String msg = e.getMessage();
 			logger.error("查询用户发生异常：");
 			e.printStackTrace();
 			result.setResult(false);
 			result.setMsg(msg);
-			return result;
 		}
+		return result;
     }
     
     /**
-     * 查询菜单
-     * @return
-     * String
+     * 查询分页数据
+     * @param Integer page
+     * @param Integer limit
+     * @param String username
+     * @return PageData<User>
      */
     @GetMapping(value="page")
     public PageData<User> page(
@@ -139,6 +223,9 @@ public class UserController extends BaseController {
             @RequestParam(value="queryKey", required=false) String username
             ) {
         PageImpl pageImpl = new PageImpl(page, limit);
+		if(StringUtils.isNotBlank(username)) {
+			username = "%" + username + "%";
+		}
         Page<User> p = userService.pageList(pageImpl, username);
         PageData<User> pageData = new PageData<>();
         pageData.setCode(0);
@@ -150,9 +237,8 @@ public class UserController extends BaseController {
     
     /**
      * 用户授权
-     * @param sysUser
-     * @return
-     * String
+     * @param UserRoleVo userRoleVo
+     * @return ApiResult<User>
      */
     @PostMapping(value="grant")
     public ApiResult<User> grant(@RequestBody UserRoleVo userRoleVo) {
